Fix istio project level routing issue
