+++ app/application/route.js
@@ -151,6 +151,7 @@ export default Route.extend({
       access.clearToken().finally(() => {
         get(this, 'tab-session').clear();
         set(this, `session.${ C.SESSION.CONTAINER_ROUTE }`, undefined);
+        set(this, `session.${ C.SESSION.ISTIO_ROUTE }`, undefined);
         set(this, `session.${ C.SESSION.CLUSTER_ROUTE }`, undefined);
         set(this, `session.${ C.SESSION.PROJECT_ROUTE }`, undefined);
