+++ lib/istio/addon/components/istio-nav/template.hbs
@@ -2,7 +2,7 @@
   <ul class="tab-nav">
     {{#if owner}}
       <li>
-        {{#link-to "index"}}{{t "istio.nav.graph"}}{{/link-to}}
+        {{#link-to "graph"}}{{t "istio.nav.graph"}}{{/link-to}}
       </li>
     {{/if}}
     <li>
