+++ lib/istio/addon/metrics/route.js
@@ -8,6 +8,7 @@ export default Route.extend({
   session:     service(),
 
   setDefaultRoute: on('activate', function() {
+    set(this, `session.${ C.SESSION.ISTIO_ROUTE }`, 'metrics');
     set(this, `session.${ C.SESSION.PROJECT_ROUTE }`, 'authenticated.project.istio.metrics');
   }),
 });
