+++ lib/istio/addon/rules/route.js
@@ -29,6 +29,7 @@ export default Route.extend({
   },
 
   setDefaultRoute: on('activate', function() {
+    set(this, `session.${ C.SESSION.ISTIO_ROUTE }`, 'rules');
     set(this, `session.${ C.SESSION.PROJECT_ROUTE }`, 'authenticated.project.istio.rules');
   }),
 });
