+++ lib/shared/addon/utils/constants.js
@@ -459,6 +459,7 @@ var C = {
     // @TODO-2.0 remove most of these
     BACK_TO:         'backTo',
     CONTAINER_ROUTE: 'containerSubRoute',
+    ISTIO_ROUTE:     'istioSubRoute',
     PROJECT_ROUTE:   'projectRoute',
     CLUSTER_ROUTE:   'clusterRoute',
     GITHUB_CACHE:    'githubCache',
